split SSL classes into their own files
